requirements.txt
pillow==11.0.0
pyperclip==1.9.0
pystray==0.19.5
six==1.17.0
